Extract style from Grid into its own JS file.
